# lib/api-zod/src/generated/types/patrolRoute.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Police Smart System API
 * OpenAPI spec version: 1.0.0
 */
import type { PatrolRouteStartPoint } from "./patrolRouteStartPoint";
import type { PatrolRouteWaypointsItem } from "./patrolRouteWaypointsItem";

export interface PatrolRoute {
  id: number;
  officerId: number;
  officerName: string;
  startPoint: PatrolRouteStartPoint;
  waypoints: PatrolRouteWaypointsItem[];
  estimatedDuration: number;
  coverageArea?: string;
  riskZones?: string[];
}