# lib/api-zod/src/generated/types/patrolRoutesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Police Smart System API
 * OpenAPI spec version: 1.0.0
 */
import type { PatrolRoute } from "./patrolRoute";

export interface PatrolRoutesResponse {
  routes: PatrolRoute[];
  totalOfficers: number;
}